new: completely overworked WIDE parts of FIsH
In order to be more easy portable all variables users may need to adjust are moved to FIsH.porting
while FIsH.me contains global variables for FIsH a user/dev normally would NOT change.

The log has been enriched with some more useful information and the stopping of android services
is now more aggressive. Apropos aggressive: now we use fuser to kill ALL processes before trying
to unmount a disk (excluding init and ourself).

Another porting stuff was added: we extract the FISHFOOD but WITHOUT a possibly existing file
/init. When you prepare your own ramdisk it may or may not contain a file /init but we do
not want to overwrite the current running /init by just cpio yet.
The reason is it makes a difference to use cp or mv or cpio > to overwrite.
The short version: using mv is the best choice here. So we extract first without /init (if there is
one) and then extract /init ONLY but to a different directory. Afterwards we use mv to overwrite the
current /init with either an extracted /init or your own init name configured by the new FOODBIN
variable.

Now FIsH also
- checks the exec bit before it tries to start it..
- save the current SELinux mode
- restores the saved SELinux mode when something goes wrong
- umounts when a mount point is mounted only

Besides that this commit introduces callmeFIsH a caller script which gets executed at very first.
The only task callmeFIsH has is to prepare the whole FIsH to get started out of /system and then
starting FIsH from /res. After this it immediately exists to not keep open tasks on /system.

Also new are some parts of the installer e.g. starting adb server first and then wait for a
connected device before continue the installation. The new installer keeping track of an older
FIsH install and removes the previous one before as well.