new: my own PS implementation
found several issues due to some debugging and trying to
improve the speed - in short: the busybox implementation is buggy.
this commit introduces F_PS function which just works.

in more detail:
When using "ps -o pids,args" it will just not display the PIDs - only args.
When using "ps -o args,pids" it will display the PIDs but busybox grep
doesn't work on it when using in busybox sh. So I dunno which busybox
target is to blame but its busybox related in any case.
This commit adds a new function "F_PS" which is a custom
implementation of the ps command based on what we can find in /proc.

Change-Id: I39f30c0d880bf234886034c46823e95c52e9eaf2