new: closes #1 and speedup kill & umount
the processes to be excluded from the global kill can be
chosen by the new variables in fishing/FIsH.porting:
EXCLUDEPROCS / EXCLUDEPIDS

This commit makes heavy advantage of the new F_PS
function introduced in a previous commit and enhances the
killing in any possible way to speed up as much as we can.

Previous runs took about 10-15 sec to kill processes which is
about 3-5 sec now (on my device at least).
Instead of killing pid by pid in a slow for loop we create a list
of pids, exclude those we do not want to kill (see above vars)
and then using one single kill cmd to blow them all.

Change-Id: I31cee335452816e39e4aea02cea29f2ccff34aa8